groove/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.6
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.4.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
